refactor(seeders): carve reference data into seeders/reference/
Permissions, roles, role-permissions, email templates, and application
questions are reference data — idempotent and safe in any environment.
Move them into seeders/reference/ with a runReferenceSeeders entry
point gated by assertSeedAllowed('reference'), which permits all four
NODE_ENVs (including production).

The legacy runAllSeeders flow keeps the same call order via re-exports
from the new index.

Per design doc step 3.
